Correct data size for  IPV6_V6ONLY

+++ src/socket.cpp
@@ -114,7 +114,7 @@ void CSocket::Init ( const quint16 iNewPortNumber, const quint16 iNewQosNumber, 
 
         // The IPV6_V6ONLY socket option must be false in order for the socket to listen on both protocols.
         // On Linux it's false by default on most (all?) distros, but on Windows it is true by default
-        const uint8_t no = 0;
+        const int no = 0;
         if ( setsockopt ( UdpSocket, IPPROTO_IPV6, IPV6_V6ONLY, (const char*) &no, sizeof ( no ) ) == -1 )
         {
             throw CGenErr ( "setsockopt for IPV6_V6ONLY failed", "Network Error" );
